+++ Rpc/ClientInterface.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace ManaPHP\Rpc;
+
+interface ClientInterface
+{
+    /**
+     * @param string $method
+     * @param array  $params
+     * @param array  $options
+     *
+     * @return mixed
+     */
+    public function invoke($method, $params = [], $options = []);
+}
